Move middleware to featureflag package
